fixing broken

+++ smacc_client_library/move_base_z_client/custom_planners/backward_global_planner/src/backward_global_planner.cpp
@@ -179,7 +179,7 @@ bool BackwardGlobalPlanner::makePlan(const geometry_msgs::PoseStamped &start,
     costmap_2d::Costmap2D *costmap2d = this->costmap_ros_->getCostmap();
     for (auto &p : plan)
     {
-        uint64_t mx, my;
+        unsigned int mx, my;
         costmap2d->worldToMap(p.pose.position.x, p.pose.position.y, mx, my);
         auto cost = costmap2d->getCost(mx, my);
